Moved some endpoints
