+++ src/WebServerManager.hpp
@@ -12,8 +12,8 @@
 #include "EmotionState.hpp"
 #include "FanController.hpp"
 #include "TiltController.hpp"
-#include "WebEndpoints/Devices/EarsEndpoint.hpp"
-#include "WebEndpoints/Devices/FanEndpoint.hpp"
+#include "WebEndpoints/Ears/EarsEndpoint.hpp"
+#include "WebEndpoints/Fan/FanEndpoint.hpp"
 #include "WebEndpoints/Emotions/EmotionEndpoint.hpp"
 #include "WebEndpoints/Emotions/EmotionsEndpoint.hpp"
 #include "WebEndpoints/Files/FileEndpoint.hpp"
